setup to move Add Record functionality to new stack/scroll view
